Rename hooks file to deviceHooks
