+++ src/components/MenuBar/DeviceSelector/VideoInputList/VideoInputList.tsx
@@ -3,7 +3,7 @@ import { FormControl, MenuItem, Typography, Select } from '@material-ui/core';
 import { LocalVideoTrack } from 'twilio-video';
 import { makeStyles } from '@material-ui/core/styles';
 import useVideoContext from '../../../../hooks/useVideoContext/useVideoContext';
-import { useVideoInputDevices } from '../hooks/hooks';
+import { useVideoInputDevices } from '../deviceHooks/deviceHooks';
 import VideoTrack from '../../../VideoTrack/VideoTrack';
 
 const useStyles = makeStyles({
