+++ src/components/MenuBar/DeviceSelector/deviceHooks/deviceHooks.tsx
